+++ aws-sns-lambda/test/.gitignore
@@ -1,0 +1,1 @@
+hello.zip
