+++ aws-sns-lambda/test/hello.py
@@ -1,0 +1,11 @@
+
+def hello(event, context):
+    message = 'Hello {}!'.format(event['name'])
+    print(message)
+    return {
+        "message": message,
+    }
+
+
+if __name__ == '__main__':
+    print(hello({'name': 'Lambda'}, {}))
